+++ papermerge_cli/exceptions.py
@@ -1,0 +1,2 @@
+class FileMimeTypeUnknown(Exception):
+    """Raised when mime of type to be uploaded cannot be guessed"""