+++ papermerge_cli/rest/__init__.py
@@ -2,6 +2,7 @@ from .documents import upload as upload_document
 from .nodes import (create_folder, get_nodes, node_add_tags, node_assign_tags,
                     node_remove_tags)
 from .users import get_me
+from .version import get_server_version
 
 __all__ = [
     get_me,
@@ -10,5 +11,6 @@ __all__ = [
     node_remove_tags,
     node_assign_tags,
     upload_document,
-    create_folder
+    create_folder,
+    get_server_version
 ]